Update urg_library to ver.1.2.10
The pinned revision was 27 commits behind upstream. Moving to ver.1.2.10
picks up the missing <unistd.h> include for usleep(), so the force-include
workaround added for GCC 14 and later can go away.

ver.1.2.10 changes the C API from int and long to fixed-width integer types,
so match the node to the new signatures:

- distance_ becomes std::vector<int32_t>, as urg_get_distance() and friends
  now take int32_t data[].
- The time_stamp locals and get_synchronized_time()'s parameter become
  int32_t, following urg_get_distance() and urg_time_stamp().
- min_dis and max_dis become int32_t, following urg_distance_min_max().
- last_hardware_time_stamp_ becomes int32_t, since it only ever holds a
  sensor time stamp.

get_time_stamp_offset()'s system_time_stamp is now int64_t. It holds the
result of rclcpp::Time::nanoseconds() rather than anything from urg_library,
and long would truncate it on platforms where it is 32 bits.

The bump also brings upstream fixes to urg_serial_find_port(), urg_sleep(),
send_distance_command() and the tcpclient non-blocking read, and relicenses
the C++ API from LGPL to the Simplified BSD License. The C API this package
uses was already Simplified BSD, so the license note in the READMEs is
unchanged.

Verified with a full build of urg_node2, urg_node2_node and urg_node2_test
in the ros:lyrical-ros-base container (Ubuntu 26.04, CMake 4.2.3, GCC 15).

+++ include/urg_node2/urg_node2.hpp
@@ -214,7 +214,7 @@ private:
    * @param[in] system_time_stamp システム時刻
    * @return 補正後タイムスタンプ
    */
-  rclcpp::Time get_synchronized_time(long time_stamp, rclcpp::Time system_time_stamp);
+  rclcpp::Time get_synchronized_time(int32_t time_stamp, rclcpp::Time system_time_stamp);
 
   /**
    * @brief 強度モード対応確認
@@ -373,7 +373,7 @@ private:
   /** 同期モード用変数 */
   double hardware_clock_;
   /** 同期モード用変数 */
-  long int last_hardware_time_stamp_;
+  int32_t last_hardware_time_stamp_;
   /** 同期モード用変数 */
   double hardware_clock_adj_;
   /** 同期モード用変数 */
@@ -387,7 +387,7 @@ private:
   int last_step_;
 
   /** スキャンデータ受信領域 */
-  std::vector<long> distance_;
+  std::vector<int32_t> distance_;
   /** 強度データ受信領域 */
   std::vector<unsigned short> intensity_;
